+++ test/dab-setOptions.js
@@ -6,27 +6,24 @@ const chai = require('chai'),
 
 chai.use(chaiSubset)
 
-const Cls = require('../index')
+const Cls = require('../dab')
 
-describe('setOptions', function () {
+describe('Dab - setOptions', function () {
   it('should return the default options', function () {
     const cls = new Cls()
     expect(cls.options).to.eql({
       limit: 25,
-      ns: 'default',
       options: {}
     })
   })
 
   it('should return the merged options', function () {
     const cls = new Cls({
-      ns: 'test',
       key1: 'key1',
       key2: 'key2'
     })
     expect(cls.options).to.eql({
       limit: 25,
-      ns: 'test',
       key1: 'key1',
       key2: 'key2',      
       options: {}